requirements.txt
Flask==2.1.2
python-telegram-bot==13.15